genes/pombe/Epe1/Epe1-bioinformatics/pyproject.toml
[project]
name = "epe1-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]
